+++ scripts/install.sh
@@ -1,39 +1,39 @@
 #!/bin/bash
 
 #echo '=== MariaDB info ==='
-#while ! [ "${LDSIDBPASS}" = "${LDSIDBPASS1}" ]; do
+#while ! [ "${LOGESPDBPASS}" = "${LOGESPDBPASS1}" ]; do
 #    read -sp 'MariaDB Root Password: ' MDBPASS
 #    read -sp 'MariaDB Root Password (again): ' MDBPASS1
 #done
 
 
-echo '=== LDSI Database Info ==='
-read -p 'DB Instance Name: ' LDSIDBINSTANCE
-read -p 'DB Username: ' LDSIDBUSER
-LDSIDBPASS='1'
-LDSIDBPASS1='1'
-while ! [ "${LDSIDBPASS}" = "${LDSIDBPASS1}" ]; do
-    read -sp 'DB Password: ' LDSIDBPASS
-    read -sp 'DB Password (again): ' LDSIDBPASS1
+echo '=== LogESP Database Info ==='
+read -p 'DB Instance Name: ' LOGESPDBINSTANCE
+read -p 'DB Username: ' LOGESPDBUSER
+LOGESPDBPASS='1'
+LOGESPDBPASS1='1'
+while ! [ "${LOGESPDBPASS}" = "${LOGESPDBPASS1}" ]; do
+    read -sp 'DB Password: ' LOGESPDBPASS
+    read -sp 'DB Password (again): ' LOGESPDBPASS1
 done
 
-if [ "${LDSIDBPASS}" = "${MDBPASS}" ]; then
-    echo "LDSI DB password should not be the MariaDB root password! Exiting."
+if [ "${LOGESPDBPASS}" = "${MDBPASS}" ]; then
+    echo "LogESP DB password should not be the MariaDB root password! Exiting."
     exit 1
 fi
 
 echo '=== Server Info ==='
-read -p 'Server FQDN (or IP): ' LDSISERVERFQDN
+read -p 'Server FQDN (or IP): ' LOGESPSERVERFQDN
 
 # Update apt, install dependencies
 
 # Create mysql database, account
 
-# Create virtual env, get LDSI
+# Create virtual env, get LogESP
 
 # Copy syslog config files, link nginx config, etc
 
-# Set server name in nginx conf/ldsi settings
+# Set server name in nginx conf/logesp settings
 
 # Put lines in /etc/rc.local